refactor MainActivity into HomeActivity and MainActivity. - this separates the first time launch logic and the home screen logic - this also removes the weird animation when going back to the home screen. My best guess is that android will play a special animation for only the MainActivity.
