+++ app/src/main/java/edu/uw/ischool/opensecrets/EntryOverviewEditActivity.kt
@@ -43,7 +43,7 @@ class EntryOverviewEditActivity : AppCompatActivity() {
             startActivity(
                 Intent(
                     this,
-                    MainActivity::class.java
+                    HomeActivity::class.java
                 ).addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
                     .addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
             )
