+++ app/src/main/java/edu/uw/ischool/opensecrets/EntryTextActivity.kt
@@ -64,7 +64,7 @@ class EntryTextActivity : AppCompatActivity() {
             startActivity(
                 Intent(
                     this,
-                    MainActivity::class.java
+                    HomeActivity::class.java
                 ).addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
                     .addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
             )
