+++ app/src/main/java/edu/uw/ischool/opensecrets/OptionActivity.kt
@@ -37,7 +37,7 @@ class OptionActivity : AppCompatActivity() {
             startActivity(
                 Intent(
                     this,
-                    MainActivity::class.java
+                    HomeActivity::class.java
                 ).addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
                     .addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
             )
