+++ app/src/main/java/edu/uw/ischool/opensecrets/SearchActivity.kt
@@ -26,7 +26,7 @@ class SearchActivity : AppCompatActivity() {
             startActivity(
                 Intent(
                     this,
-                    MainActivity::class.java
+                    HomeActivity::class.java
                 ).addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
                     .addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
             )
